Refactorisation - consistent package structure
Move all orm entity classes into ..symphony.entity package.
Move all service classes into ..symphony.service package.
Move all HTTP endpoints into ..symphony.web package.
Separate "Normalizer" classes and make them public.

Also add \@TypeDef annotations for array types and Json in entity classes, where needed.
