+++ symphony-ws/src/main/java/se/havochvatten/symphony/entity/CalculationArea.java
@@ -2,7 +2,6 @@ package se.havochvatten.symphony.entity;
 
 import org.hibernate.annotations.Fetch;
 import org.hibernate.annotations.FetchMode;
-import se.havochvatten.symphony.scenario.ScenarioArea;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
@@ -10,9 +9,7 @@ import javax.validation.constraints.Size;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlTransient;
 import java.io.Serializable;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Set;
 
 @Entity
 @Table(name = "calculationarea")
