+++ symphony-ws/src/main/java/se/havochvatten/symphony/service/CalculationAreaService.java
@@ -14,8 +14,8 @@ import se.havochvatten.symphony.exception.SymphonyStandardAppException;
 import se.havochvatten.symphony.mapper.AreaSelectionResponseDtoMapper;
 import se.havochvatten.symphony.dto.AreaSelectionResponseDto.AreaOverlapFragment;
 import se.havochvatten.symphony.mapper.CalculationAreaMapper;
-import se.havochvatten.symphony.scenario.Scenario;
-import se.havochvatten.symphony.scenario.ScenarioArea;
+import se.havochvatten.symphony.entity.Scenario;
+import se.havochvatten.symphony.entity.ScenarioArea;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
