+++ symphony-ws/src/main/java/se/havochvatten/symphony/service/normalizer/DomainNormalizer.java
@@ -1,0 +1,4 @@
+package se.havochvatten.symphony.service.normalizer;
+
+class DomainNormalizer extends RasterNormalizer {
+}
