+++ symphony-ws/src/main/java/se/havochvatten/symphony/service/normalizer/UserDefinedValueNormalizer.java
@@ -1,0 +1,4 @@
+package se.havochvatten.symphony.service.normalizer;
+
+class UserDefinedValueNormalizer extends RasterNormalizer {
+}
